requirements.txt
requests==2.31.0
python-dotenv==1.0.0
jsonschema==4.17.3
tqdm==4.65.0